backend/requirements.txt
# API Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
SQLAlchemy==2.0.41
aiosqlite==0.20.0

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Document Processing
PyMuPDF==1.23.26  # (fitz)
pdfplumber==0.11.6
python-docx==1.1.2
lxml==5.4.0

# Vector Store
weaviate-client==3.26.7

# LLM & Embeddings
langchain==0.3.25
langchain-community==0.3.24
langchain-ollama==0.3.3

# ML & Transformers
torch==2.6.0
transformers==4.51.2

# Utilities
Markdown==3.8
pydantic==2.11.4
